requirements.txt
asgiref==3.7.2
bleach==6.0.0
boto3==1.28.53
botocore==1.31.53
dj-database-url==2.1.0
Django==4.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
django-storages==1.14
flake8==6.1.0
jmespath==1.0.1
mccabe==0.7.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
webencodings==0.5.1
whitenoise==6.5.0
